@@ -1,41 +1,63 @@
 name: chit_chat
 description: "A new Flutter project."
 publish_to: 'none'
-version: 1.0.0
+version: 1.2.1
 
 environment:
-  sdk: '>=3.3.3 <4.0.0'
+  sdk: ^3.5.1
 
 dependencies:
-  cloud_firestore: ^5.0.2
-  connectivity_plus: ^6.0.2
-  firebase_auth: ^5.1.1
-  firebase_core: ^3.1.1
-  firebase_messaging: 15.0.2
+  animated_theme_switcher: ^2.0.10
+  animations: ^2.0.11
+  audio_waveforms: ^1.2.0
+  cached_network_image: ^3.4.1
+  cloud_firestore: ^5.5.0
+  connectivity_plus: ^6.0.5
+  dio: ^5.7.0
+  firebase_auth: ^5.3.3
+  firebase_core: ^3.8.0
+  firebase_messaging: ^15.2.4
+  firebase_storage: ^12.3.6
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.6
+  flutter_animate: ^4.5.2
+  flutter_bloc: ^9.0.0
+  flutter_cache_manager: ^3.4.1
+  flutter_local_notifications: ^19.0.0
   flutter_svg: ^2.0.10+1
-  googleapis_auth: ^1.6.0
-  http: ^1.2.1
-  image_picker: ^1.1.0
-  intl: ^0.19.0
-  lottie: ^3.1.0 
-  rive: ^0.13.1
-  flutter_local_notifications: ^17.1.0
-  cached_network_image: ^3.3.1
-  firebase_storage: ^12.1.0
-  animations: ^2.0.11
-  shared_preferences: ^2.2.3
   gap: ^3.0.1
+  googleapis_auth: ^2.0.0
+  http: ^1.2.2
+  image_picker: ^1.1.2
+  image_size_getter: ^2.4.0
+  intl: ^0.20.1
+  lottie: ^3.1.2
+  package_info_plus: ^8.0.2
+  open_file: ^3.5.10
+  path:
+  path_provider: ^2.1.5
+  photo_view: ^0.15.0
   popover: ^0.3.0+1
- 
+  record: ^6.0.0
+  share_plus: ^10.1.2
+  shared_preferences: ^2.3.2
+  timer_count_down: ^2.2.2
+  video_player: ^2.9.1
+  video_thumbnail: ^0.5.3
+  permission_handler: ^11.4.0
+  media_store_plus: ^0.1.3
 
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  flutter_lints: ^5.0.0
+  flutter_native_splash: 
+  flutter_launcher_icons: 
 
 flutter_native_splash:
   web: false
   image: assets/icon/splash_icon.png
-  # background_image: assets/images/bg-login.jpg
   color: "#d6e2ea"
   android: true
 
@@ -43,36 +65,28 @@ flutter_native_splash:
     image: assets/icon/splash_icon.png
     color: "#d6e2ea"
 
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^4.0.0
-  flutter_native_splash:
-  flutter_launcher_icons: ^0.13.1
-
-
-
 flutter_launcher_icons:
   android: "launcher_icon"
   ios: true
-  image_path: "assets/icon/app_icon.png"
+  image_path: "assets/Icon/app_icon.png"
+  adaptive_icon_foreground: "assets/Icon/app_icon.png"
+  adaptive_icon_background: "#ffffff"
+  remove_alpha_ios: true
  
-flutter:
-  uses-material-design: true
 
-  assets:
+flutter:
+ uses-material-design: true
+ assets:
     - assets/
     - assets/images/
-    - assets/rivfiles/
     - assets/svg/
     - assets/lottie/
 
-  fonts:
+ fonts:
     - family: "Roboto-Regular"
       fonts:
         - asset: fonts/Roboto/Roboto-Regular.ttf
-
+        
     - family: "Roboto-Bold"
       fonts:
         - asset: fonts/Roboto/Roboto-Bold.ttf
@@ -80,10 +94,3 @@ flutter:
     - family: "Roboto-Medium"
       fonts:
         - asset: fonts/Roboto/Roboto-Medium.ttf
-
-    - family: "Roboto-Light"
-      fonts:
-        - asset: fonts/Roboto/Roboto-Light.ttf
-    - family: "Roboto-B"
-      fonts:
-        - asset: fonts/Roboto/Roboto-Light.ttf
