@@ -1,11 +1,11 @@
 import 'package:chit_chat/model/user_data.dart';
-import 'package:chit_chat/model/user_model.dart';
-import 'package:chit_chat/utils/util.dart';
+import 'package:chit_chat/res/common_instants.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+
 part 'auth_state.dart';
 
 class AuthCubit extends Cubit<AuthState> {
@@ -13,27 +13,23 @@ class AuthCubit extends Cubit<AuthState> {
       : super(
           AuthViewState(
               status: PageStatus.notSignedIn,
-              userModel: UserModel(email: '', password: ''),
+              userModel: UserData(userEmail: '', password: ''),
               buttonLoader: false),
         );
 
   FirebaseFirestore firebaseFirestore = FirebaseFirestore.instance;
-
-  FirebaseAuth firebaseAuth = FirebaseAuth.instance;
-
   FirebaseMessaging firebaseMessaging = FirebaseMessaging.instance;
 
-  Util util = Util();
-  onInit() {
+  loginInizialize() {
     emit(AuthViewState(
         status: PageStatus.notSignedIn,
-        userModel: UserModel(email: '', password: ''),
+        userModel: UserData(userEmail: '', password: ''),
         buttonLoader: false));
   }
 
-  doSignIn(PageStatus status, UserModel userModel) {
+  void doSignIn(PageStatus status, UserData userModel) {
     if (status == PageStatus.signIn) {
-      if (userModel.email == null || userModel.email!.isEmpty) {
+      if (userModel.userEmail == null || userModel.userEmail!.isEmpty) {
         emit(AuthToast(type: 'error', text: 'Please enter email address'));
       } else if (userModel.password == null ||
           userModel.password!.isEmpty ||
@@ -43,33 +39,27 @@ class AuthCubit extends Cubit<AuthState> {
         util.checkNetwork().then((value) async {
           try {
             emit(AuthLoading());
-
             await firebaseAuth.signInWithEmailAndPassword(
-                email: userModel.email!, password: userModel.password!);
+                email: userModel.userEmail!, password: userModel.password!);
 
-            await firebaseMessaging.getToken().then((token) async {
-              await firebaseFirestore
-                  .collection('users')
-                  .where('uid', isEqualTo: firebaseAuth.currentUser!.uid)
-                  .get()
-                  .then((user) {
-                final UserData userData =
-                    UserData.fromJson(user.docs[0].data());
-                DocumentReference messageRef =
-                    firebaseFirestore.collection('users').doc(user.docs[0].id);
-                if (userData.fCM != token) {
-                  userData.fCM = token;
-                  messageRef.update(userData.toJson());
-                }
-              });
-            });
+            //refreshing instance
+            firebaseAuth.currentUser?.reload();
 
             emit(AuthCancelLoading());
-            emit(AuthUserLoginSuccess());
+
+            if (firebaseAuth.currentUser?.emailVerified == true) {
+              emit(AuthUserLoginSuccess());
+            } else {
+              emit(AuthVerifyUserEmail());
+            }
           } on FirebaseAuthException catch (e) {
             emit(AuthCancelLoading());
             showFirebaseError(e);
           } catch (e) {
+            emit(AuthAlert(
+                type: 'network',
+                text: 'No internet\nPlease connect to internet'));
+
             emit(AuthCancelLoading());
           }
         }).catchError((e) {
@@ -82,14 +72,14 @@ class AuthCubit extends Cubit<AuthState> {
       emit(AuthViewState(
           buttonLoader: true,
           status: PageStatus.signIn,
-          userModel: UserModel(email: '', password: '')));
+          userModel: UserData(userEmail: '', password: '')));
     }
   }
 
-  dosignUP(UserModel userModel) async {
-    if (userModel.name == null || userModel.name!.isEmpty) {
+  dosignUP(UserData userModel) async {
+    if (userModel.userName == null || userModel.userName!.isEmpty) {
       emit(AuthToast(type: 'error', text: 'Please enter your name'));
-    } else if (userModel.email == null || userModel.email!.isEmpty) {
+    } else if (userModel.userEmail == null || userModel.userEmail!.isEmpty) {
       emit(AuthToast(type: 'error', text: 'Please enter email address'));
     } else if (userModel.password == null ||
         userModel.password!.isEmpty ||
@@ -99,17 +89,25 @@ class AuthCubit extends Cubit<AuthState> {
       util.checkNetwork().then((value) async {
         try {
           emit(AuthLoading());
+
+          if (firebaseAuth.currentUser != null) {
+            firebaseAuth.signOut();
+          }
+
           //sign up user
           await firebaseAuth
               .createUserWithEmailAndPassword(
-                  email: userModel.email!, password: userModel.password!)
+                  email: userModel.userEmail!, password: userModel.password!)
               .then((value) async {
+            firebaseAuth = FirebaseAuth.instance;
+
             //update username
-            await firebaseAuth.currentUser!.updateDisplayName(userModel.name);
+            await firebaseAuth.currentUser!
+                .updateDisplayName(userModel.userName);
 
             final UserData user = UserData(
-              userEmail: userModel.email,
-              userName: userModel.name,
+              userEmail: userModel.userEmail,
+              userName: userModel.userName,
               uid: firebaseAuth.currentUser!.uid,
             );
 
@@ -120,7 +118,7 @@ class AuthCubit extends Cubit<AuthState> {
             emit(AuthViewState(
                 buttonLoader: false,
                 status: PageStatus.signIn,
-                userModel: UserModel(email: '', password: '')));
+                userModel: UserData(userEmail: '', password: '')));
           });
         } on FirebaseAuthException catch (e) {
           emit(AuthCancelLoading());
@@ -128,24 +126,6 @@ class AuthCubit extends Cubit<AuthState> {
         }
       }).catchError((e) {
         emit(AuthAlert(type: 'network', text: 'Please connect to internet'));
-      });
-    }
-  }
-
-  verifyEmail(String email) {
-    // firebaseAuth.;
-  }
-
-  forgotPassword(String email) {
-    if (email.isEmpty) {
-      emit(AuthToast(type: 'error', text: 'Please enter email'));
-    } else {
-      util.checkNetwork().then((value) async {
-        firebaseAuth.sendPasswordResetEmail(email: email).then((value) {
-          emit(AuthPasswordResetMailSent());
-        }).catchError((e) {
-          showFirebaseError(e);
-        });
       });
     }
   }
@@ -174,6 +154,37 @@ class AuthCubit extends Cubit<AuthState> {
     emit(AuthViewState(
         buttonLoader: true,
         status: PageStatus.notSignedIn,
-        userModel: UserModel(email: '', password: '')));
+        userModel: UserData(userEmail: '', password: '')));
+  }
+
+  onInit() {
+    emit(AuthViewState(
+        status: PageStatus.notSignedIn,
+        userModel: UserData(userEmail: '', password: ''),
+        buttonLoader: false));
+  }
+
+  verifyEmail(BuildContext context) {
+    firebaseAuth.currentUser!.sendEmailVerification().catchError((e) {
+      if (e.code == 'too-many-requests' && context.mounted) {
+        util.doAlert(context, 'Too Many Request \nPlease try later', 'error');
+      } else {
+        showFirebaseError(e);
+      }
+    });
+  }
+
+  forgotPassword(String email) {
+    if (email.isEmpty) {
+      emit(AuthToast(type: 'error', text: 'Please enter email'));
+    } else {
+      util.checkNetwork().then((value) async {
+        firebaseAuth.sendPasswordResetEmail(email: email).then((value) {
+          emit(AuthPasswordResetMailSent());
+        }).catchError((e) {
+          showFirebaseError(e);
+        });
+      });
+    }
   }
 }
