@@ -1,7 +1,15 @@
-class UserModel {
-  String? email;
-  String? password;
-  String? name;
-  String? profileURL;
-  UserModel({this.name, this.email, this.password, this.profileURL});
-}
+// class UserModel {
+//   String? email;
+//   String? password;
+//   String? name;
+//   String? profileURL;
+//   bool? isVerified;
+//   String? phoneNumber;
+//   UserModel(
+//       {this.phoneNumber,
+//       this.name,
+//       this.email,
+//       this.password,
+//       this.profileURL,
+//       this.isVerified});
+// }
