@@ -1,17 +1,21 @@
+import 'package:animated_theme_switcher/animated_theme_switcher.dart';
 import 'package:animations/animations.dart';
+import 'package:chit_chat/controller/chat_cubit/chat_cubit.dart';
 import 'package:chit_chat/controller/home_cubit/home_cubit.dart';
+import 'package:chit_chat/controller/profile_cubit/profile_cubit.dart';
+import 'package:chit_chat/controller/update_cubit/update_cubit.dart';
 import 'package:chit_chat/res/colors.dart';
-import 'package:chit_chat/res/custom_widget/loading_widget.dart';
-import 'package:chit_chat/res/fonts.dart';
+import 'package:chit_chat/res/common_instants.dart';
+import 'package:chit_chat/res/custom_widget/svg_icon.dart';
 import 'package:chit_chat/view/screen/chat_page.dart';
 import 'package:chit_chat/view/screen/search_page.dart';
 import 'package:chit_chat/view/widget/circular_profile_image.dart';
+import 'package:chit_chat/view/widget/update_dialog.dart';
+import 'package:chit_chat/view/widget/user_card.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:lottie/lottie.dart';
-import '../widget/side_menu.dart';
-import '../widget/user_card.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -20,75 +24,163 @@ class HomePage extends StatefulWidget {
   State<HomePage> createState() => _HomePageState();
 }
 
-class _HomePageState extends State<HomePage> {
+class _HomePageState extends State<HomePage> with WidgetsBindingObserver {
+  late HomeCubit _homeCubit;
   @override
   void initState() {
-    BlocProvider.of<HomeCubit>(context).onInit();
+    BlocProvider.of<HomeCubit>(context).onInit().then((v) {
+      if (mounted) {
+        BlocProvider.of<ProfileCubit>(context).getProfile();
+      }
+    });
+    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance.addPostFrameCallback((target) {
+      BlocProvider.of<HomeCubit>(context).checkNotificationStack();
+    });
     super.initState();
+  }
+
+  ThemeSwitcherState? themeSwitcherState;
+
+  @override
+  void didChangePlatformBrightness() {
+    util.changeTheme(themeSwitcherState, context);
+    super.didChangePlatformBrightness();
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    if (state == AppLifecycleState.paused ||
+        state == AppLifecycleState.inactive) {
+      _homeCubit.pauseStream();
+    } else if (state == AppLifecycleState.resumed) {
+      _homeCubit.resumeStream();
+    }
+  }
+
+  @override
+  void didChangeDependencies() {
+    _homeCubit = BlocProvider.of<HomeCubit>(context);
+    super.didChangeDependencies();
+  }
+
+  @override
+  Future<void> dispose() async {
+    WidgetsBinding.instance.removeObserver(this);
+    _homeCubit.stopStream();
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.inverseSurface,
-      drawer: const SideMenu(),
-      appBar: AppBar(
-        titleSpacing: 0,
-        title: const Text('Chitchat',
-            style: TextStyle(color: AppColor.white, fontFamily: Roboto.medium)),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        iconTheme: const IconThemeData(color: AppColor.white),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          context.read<HomeCubit>().toSearch();
-        },
-        child: const Icon(
-          Icons.message_rounded,
-          color: AppColor.white,
-        ),
-      ),
-      body: BlocListener<HomeCubit, HomeState>(
-        listenWhen: (previous, current) => current is HomeActionState,
-        listener: (context, state) {
-          listener(state, context, context.read<HomeCubit>());
-        },
-        child: BlocBuilder<HomeCubit, HomeState>(
-          buildWhen: (previous, current) => current is! HomeActionState,
-          builder: (context, state) {
-            if (state is HomeReadyState) {
-              if (state.userList.isEmpty) {
-                return WelcomeWidget(
-                  username: state.user.name ?? '',
-                );
-              }
-              return ListView.builder(
-                itemCount: state.userList.length,
-                itemBuilder: (context, index) {
-                  return OpenContainer(
-                    tappable: false,
-                    openColor: Theme.of(context).colorScheme.inverseSurface,
-                    closedColor: Theme.of(context).colorScheme.inverseSurface,
-                    openBuilder: (context, action) =>
-                        ChatPage(userData: state.userList[index]),
-                    closedBuilder: (context, action) => UserCard(
-                      user: state.userList[index],
-                      onTap: (userData) => action.call(),
-                    ),
-                  );
+    return ThemeSwitcher.switcher(
+      builder: (p0, switcher) => Builder(builder: (context) {
+        themeSwitcherState ??= switcher;
+        return Scaffold(
+          backgroundColor: Theme.of(context).colorScheme.surfaceDim,
+          appBar: AppBar(
+            toolbarHeight: kTextTabBarHeight + 30,
+            titleSpacing: 20,
+            title: Text(
+              'ChitChat',
+              style: style.text.regularLarge.copyWith(color: AppColor.darkBlue),
+            ),
+            backgroundColor: Theme.of(context).colorScheme.surfaceDim,
+            actions: [
+              BlocListener<UpdateCubit, UpdateState>(
+                listener: (context, state) {
+                  if (state is UpdateAvailableState) {
+                    util.slideInDialog(context, UpdateDialog(), false);
+                  } else if (state is NetworkErrorState) {
+                    util.doAlert(
+                        context, 'Please connect to internet', 'network');
+                  } else if (state is UptoDateState) {
+                    util.showSnackbar(context, 'You are up to date', 'success');
+                  } else if (state is UpdateAlertState) {
+                    util.doAlert(context, state.text, state.type);
+                  }
                 },
-              );
-            } else if (state is HomeChatLoading) {
-              return ListView.builder(
-                  itemCount: 3,
+                child: BlocBuilder<ProfileCubit, ProfileState>(
+                  buildWhen: (previous, current) =>
+                      current is! ProfileActionState,
+                  builder: (context, state) {
+                    if (state is ProfileInitial) {
+                      return GestureDetector(
+                        behavior: HitTestBehavior.opaque,
+                        onTap: () {
+                          Navigator.pushNamed(context, 'profile');
+                        },
+                        child: Hero(
+                          tag: 'profile',
+                          child: Container(
+                            margin: const EdgeInsets.all(10),
+                            height: 40,
+                            width: 40,
+                            child: CircularProfileImage(
+                              isNetworkImage: state.user.profileURL != null,
+                              image: state.user.profileURL,
+                            ),
+                          ),
+                        ),
+                      );
+                    } else {
+                      return const SizedBox();
+                    }
+                  },
+                ),
+              )
+            ],
+          ),
+          floatingActionButton: FloatingActionButton(
+            onPressed: () {
+              context.read<HomeCubit>().toSearch();
+            },
+            child: const SVGIcon(
+              name: 'message-plus',
+              color: AppColor.white,
+              size: 30,
+            ),
+          ),
+          body: BlocConsumer<HomeCubit, HomeState>(
+            listenWhen: (previous, current) => current is HomeActionState,
+            listener: (context, state) {
+              listener(state, context, context.read<HomeCubit>());
+            },
+            buildWhen: (previous, current) => current is! HomeActionState,
+            builder: (context, state) {
+              if (state is HomeReadyState) {
+                if (state.userList.isEmpty) {
+                  return WelcomeWidget(
+                    username: FirebaseAuth.instance.currentUser!.displayName!
+                        .toUpperCase(),
+                  );
+                }
+                return ListView.builder(
+                  itemCount: state.userList.length,
                   itemBuilder: (context, index) {
-                    return Container(
-                      decoration: BoxDecoration(
-                          border: Border(
-                              bottom: BorderSide(
-                                  width: 0.5,
-                                  color: AppColor.greyline.withOpacity(0.5)))),
-                      child: ListTile(
+                    return OpenContainer(
+                      closedElevation: 0,
+                      tappable: false,
+                      openColor: Theme.of(context).colorScheme.surfaceDim,
+                      closedColor: Theme.of(context).colorScheme.surfaceDim,
+                      openBuilder: (context, action) =>
+                          ChatPage(userData: state.userList[index]),
+                      closedBuilder: (context, action) => UserCard(
+                        user: state.userList[index],
+                        onTap: (userData) {
+                          BlocProvider.of<ChatCubit>(context)
+                              .onInit(state.userList[index].uid!, userData);
+                          action.call();
+                        },
+                      ),
+                    );
+                  },
+                );
+              } else if (state is HomeChatLoading) {
+                return ListView.builder(
+                    itemCount: 3,
+                    itemBuilder: (context, index) {
+                      return ListTile(
                           leading: Container(
                               width: 50,
                               height: 50,
@@ -113,67 +205,26 @@ class _HomePageState extends State<HomePage> {
                               width: MediaQuery.sizeOf(context).width * 0.15,
                               color: AppColor.greyline,
                             ),
-                          )),
-                    );
-                  });
-            } else {
-              return const SizedBox();
-            }
-          },
-        ),
-      ),
+                          ));
+                    });
+              } else {
+                return const SizedBox();
+              }
+            },
+          ),
+        );
+      }),
     );
   }
 
   void listener(
       HomeState state, BuildContext context, HomeCubit homeController) {
-    if (state is HomeSignOut) {
-      Navigator.pushNamedAndRemoveUntil(context, 'login', (route) => false);
-    } else if (state is HomeToSearch) {
+    if (state is HomeToSearch) {
       Navigator.push(
         context,
         MaterialPageRoute(
-          builder: (context) => UserSearch(chatList: homeController.userList),
+          builder: (context) => SearchPage(chatList: homeController.userList),
         ),
-      );
-    } else if (state is HomeEditProfile) {
-      showDialog(
-        context: context,
-        builder: (context) {
-          return Dialog(
-            clipBehavior: Clip.antiAliasWithSaveLayer,
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                ListTile(
-                  onTap: () {
-                    Navigator.pop(context);
-                    BlocProvider.of<HomeCubit>(context)
-                        .pickImage(ImageSource.camera);
-                  },
-                  leading: const Icon(
-                    Icons.camera,
-                  ),
-                  title: const Text('Capture Image'),
-                ),
-                ListTile(
-                  onTap: () {
-                    Navigator.pop(context);
-                    BlocProvider.of<HomeCubit>(context)
-                        .pickImage(ImageSource.gallery);
-                  },
-                  leading: const Icon(Icons.image),
-                  title: const Text('Pick from galary'),
-                ),
-              ],
-            ),
-          );
-        },
-      );
-    } else if (state is HomeScreenLoading) {
-      showDialog(
-        context: context,
-        builder: (context) => const LoadingScreen(),
       );
     }
   }
@@ -200,18 +251,10 @@ class WelcomeWidget extends StatelessWidget {
             ),
             Text(
               'Welcome ${username.toUpperCase()}',
-              style: const TextStyle(
-                fontFamily: Roboto.bold,
-                fontSize: AppFontSize.xl,
-              ),
+              style: style.text.boldXLarge,
             ),
-            const Text(
-              'Click the 👇 Button below to connect with your friends...',
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                fontSize: AppFontSize.sm,
-              ),
-            ),
+            Text('Click the  Button below to connect with your friends...',
+                textAlign: TextAlign.center, style: style.text.regular),
           ],
         ),
       ),
