@@ -1,16 +1,31 @@
 import 'dart:io';
-import 'package:chit_chat/controller/chat_cubit/chat_cubit.dart';
-import 'package:chit_chat/res/colors.dart';
-import 'package:chit_chat/res/fonts.dart';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class FileUploadDialog extends StatelessWidget {
-  final UploadFile state;
+import 'package:chit_chat/controller/chat_cubit/chat_cubit.dart';
+import 'package:chit_chat/controller/media_cubit/media_cubit.dart';
+import 'package:chit_chat/res/colors.dart';
+import 'package:chit_chat/res/common_instants.dart';
+import 'package:chit_chat/view/widget/video_preview.dart';
+
+class FileUploadDialog extends StatefulWidget {
   const FileUploadDialog({
     super.key,
-    required this.state,
   });
+
+  @override
+  State<FileUploadDialog> createState() => _FileUploadDialogState();
+}
+
+class _FileUploadDialogState extends State<FileUploadDialog> {
+  late final ChatCubit chatCubit;
+  @override
+  void initState() {
+    chatCubit = context.read<ChatCubit>();
+    chatCubit.emitfileUploadState();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -18,55 +33,74 @@ class FileUploadDialog extends StatelessWidget {
       buildWhen: (previous, current) => current is UploadFile,
       builder: (context, state) {
         if (state is UploadFile) {
-          return Dialog(
-            child: Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  const Text(
-                    'Are you sure want to send this file?',
-                    style: TextStyle(fontSize: AppFontSize.sm),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 16.0),
-                    child: Image.file(
-                      File(state.filePath),
-                      width: MediaQuery.sizeOf(context).width * 0.35,
-                      height: MediaQuery.sizeOf(context).width * 0.35,
+          return PopScope(
+            canPop: false,
+            child: Dialog(
+              child: Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Are you sure want to send this file?',
+                      style: style.text.regular,
                     ),
-                  ),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.end,
-                    children: [
-                      TextButton.icon(
-                        style: TextButton.styleFrom(
-                          foregroundColor: AppColor.blue,
+                    ImageCollageWidget(
+                        mediaType: state.mediaType, medialist: state.fileData),
+                    // buttons
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        if (state.fileStatus == FileStatus.preview)
+                          TextButton(
+                            onPressed: () {
+                              Navigator.pop(context);
+                            },
+                            child: Text(
+                              'Cancel',
+                              style: TextStyle(
+                                  color:
+                                      Theme.of(context).colorScheme.tertiary),
+                            ),
+                          ),
+                        TextButton.icon(
+                          style: TextButton.styleFrom(
+                            foregroundColor: AppColor.blue,
+                          ),
+                          iconAlignment: IconAlignment.end,
+                          icon: state.fileStatus == FileStatus.preview
+                              ? Icon(
+                                  Icons.upload,
+                                  size: style.icon.sm,
+                                  color: AppColor.blue,
+                                )
+                              : SizedBox(
+                                  height: 16,
+                                  width: 16,
+                                  child: CircularProgressIndicator(
+                                    color: Theme.of(context)
+                                        .colorScheme
+                                        .onSecondary,
+                                    strokeWidth: 2,
+                                  )),
+                          onPressed: state.fileStatus == FileStatus.preview
+                              ? () {
+                                  context
+                                      .read<ChatCubit>()
+                                      .uploadFileToFirebase();
+                                }
+                              : null,
+                          label: Text(
+                            state.fileStatus == FileStatus.preview
+                                ? 'upload'
+                                : 'uploading',
+                          ),
                         ),
-                        iconAlignment: IconAlignment.end,
-                        icon: state.fileStatus == FileStatus.preview
-                            ? const Icon(Icons.upload)
-                            : const SizedBox(
-                                height: 16,
-                                width: 16,
-                                child: CircularProgressIndicator(
-                                  strokeWidth: 2,
-                                )),
-                        onPressed: () {
-                          context
-                              .read<ChatCubit>()
-                              .uploadFileToFirebase(state.filePath);
-                        },
-                        label: Text(
-                          state.fileStatus == FileStatus.preview
-                              ? 'upload'
-                              : 'uploading',
-                        ),
-                      ),
-                    ],
-                  )
-                ],
+                      ],
+                    )
+                  ],
+                ),
               ),
             ),
           );
@@ -74,6 +108,58 @@ class FileUploadDialog extends StatelessWidget {
           return const SizedBox();
         }
       },
+    );
+  }
+}
+
+class ImageCollageWidget extends StatelessWidget {
+  const ImageCollageWidget(
+      {super.key, required this.mediaType, required this.medialist});
+  final MediaType mediaType;
+  final List<String> medialist;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.symmetric(vertical: style.insets.lg),
+      height: MediaQuery.sizeOf(context).width * 0.45,
+      child: ListView.builder(
+        scrollDirection: Axis.horizontal,
+        itemCount: medialist.length,
+        itemBuilder: (context, index) => Stack(
+          alignment: Alignment.topRight,
+          children: [
+            Container(
+              margin: EdgeInsets.only(right: style.insets.sm),
+              width: MediaQuery.sizeOf(context).width * 0.35,
+              height: MediaQuery.sizeOf(context).width * 0.45,
+              child: mediaType == MediaType.image
+                  ? Image.file(
+                      File(medialist[index]),
+                      fit: BoxFit.cover,
+                    )
+                  : VideoPreview(filepath: medialist[index]),
+            ),
+            Positioned(
+              right: 4,
+              child: IconButton(
+                  style: IconButton.styleFrom(
+                      visualDensity: VisualDensity.compact,
+                      backgroundColor: AppColor.black.withValues(alpha: 0.3),
+                      foregroundColor: AppColor.white),
+                  onPressed: () {
+                    if (medialist.length == 1) {
+                      Navigator.pop(context);
+                    }
+                    context
+                        .read<ChatCubit>()
+                        .deleteSelectedMedia(index, mediaType);
+                  },
+                  icon: Icon(Icons.delete, size: style.icon.xs)),
+            )
+          ],
+        ),
+      ),
     );
   }
 }
